Split core and data_model into clearer locations
